MARC Reader, Prints Fields
Reads MARC file (can read multiple MARC files condensed into one) and prints out 245, 100, and 650 fields, only subfield "a".  Code can be adjusted to add or delete any fields or subfields contained inside the MARC record.  

(I'm working on getting this to export to a .csv file for easier use, but it's just not there yet.)